refactor(probe): 拆出 ProbePanel 监控头部、图表组件、监控区块与交互 hook
